requirements.txt
requests
python-telegram-bot
simple-settings==0.19.1
ipdb
lxml==4.5.2